app/requirements.txt
flask
Pillow
numpy
opencv-python
ultralytics
gunicorn
gevent
gevent-websocket
flask_sockets
websockets
imageio[ffmpeg]
